List dependencies & Refactor code
