+++ README.md
@@ -1,5 +1,8 @@
 # NairalandScraper
 A scraper for emails on NL(Nairaland.com)
 
+## Installation
+pip install requirements.txt -r
+
 ## Usage
 python main_thread.py "nairaland link" "name of file to save to (optional)"
