+++ core/regexes.py
