+++ requirements.txt
@@ -1,0 +1,3 @@
+beautifulsoup4==4.7.1
+requests==2.22.0
+urllib3==1.25.3
